rename all show components to its correct name
